requirements.txt
fastapi~=0.120.0
uvicorn[standard]~=0.38.0
pydantic~=2.12.3

numpy~=2.0.2
pandas~=2.3.3
scikit-learn~=1.6.1
joblib~=1.5.2

jinja2~=3.1.4
python-dotenv~=1.2.1
groq~=0.33.0
